+++ test/tokens/interfaces/IERC721Extended.sol
@@ -3,9 +3,15 @@ pragma solidity ^0.8.20;
 
 import {IERC721Metadata} from "openzeppelin/token/ERC721/extensions/IERC721Metadata.sol";
 import {IERC721Enumerable} from "openzeppelin/token/ERC721/extensions/IERC721Enumerable.sol";
+import {IERC5267} from "openzeppelin/interfaces/IERC5267.sol";
 import {IERC4494} from "./IERC4494.sol";
 
-interface IERC721Extended is IERC721Metadata, IERC721Enumerable, IERC4494 {
+interface IERC721Extended is
+    IERC721Metadata,
+    IERC721Enumerable,
+    IERC4494,
+    IERC5267
+{
     function burn(uint256 tokenId) external;
 
     function is_minter(address minter) external view returns (bool);