+++ test/utils/interfaces/IEIP712DomainSeparator.sol
@@ -1,7 +1,9 @@
 // SPDX-License-Identifier: WTFPL
 pragma solidity ^0.8.20;
 
-interface IEIP712DomainSeparator {
+import {IERC5267} from "openzeppelin/interfaces/IERC5267.sol";
+
+interface IEIP712DomainSeparator is IERC5267 {
     function domain_separator_v4() external view returns (bytes32);
 
     function hash_typed_data_v4(