+++ test/utils/mocks/EtherReceiver.sol
@@ -5,7 +5,7 @@ pragma solidity ^0.8.20;
  * @title EtherReceiver
  * @author pcaversaccio
  * @notice Forked and adjusted accordingly from here:
- * https://github.com/mds1/multicall/blob/master/src/test/mocks/EtherSink.sol.
+ * https://github.com/mds1/multicall/blob/main/src/test/mocks/EtherSink.sol.
  * @dev Allows to test receiving ether via low-level calls.
  */
 contract EtherReceiver {